pyproject.toml
[project]
name = "gpt-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi[standard]>=0.115.6",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
]
